I have completed refactoring.
1. All classes are now in separate files. The filename is same as my class name.
2. I have created a new folder called TPC which contains all my third-person camera classes .
3 I have added a namespace TPC for all TPC camera implementations.
