+++ UnityProject/Assets/Scripts/GameConstants.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+public static class GameConstants
+{
+  public static Vector3 CameraAngleOffset { get; set; }
+  public static Vector3 CameraPositionOffset { get; set; }
+  public static float Damping { get; set; }
+  //public static float RotationSpeed { get; set; }
+  //public static float MinPitch { get; set; }
+  //public static float MaxPitch { get; set; }
+}
