+++ UnityProject/Assets/Scripts/GameConstants.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce359a7dd0999f142953d7c4963366ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
