+++ UnityProject/Assets/Scripts/TPC/Readme.txt
@@ -1,0 +1,3 @@
+This folder comprises all my third-person camera classes.
+///
+
