+++ UnityProject/Assets/Scripts/TPC/Readme.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f5370c42d7c8e404c92ca55c4b5e50ce
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
