+++ UnityProject/Assets/Scripts/TPC/TPCBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55f453fb02a1be644bd84e0886a1f58c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
