+++ UnityProject/Assets/Scripts/TPC/TPCFollow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb868f7ab811b944e9690791cf2fce90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
