+++ UnityProject/Assets/Scripts/TPC/TPCFollowTrackPosition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13aa928f379f7c441857f3ff620d9f72
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
