+++ UnityProject/Assets/Scripts/TPC/TPCFollowTrackPositionAndRotation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: febe78406935f7e498fa889402ef5577
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
