+++ UnityProject/Assets/Scripts/TPC/TPCTrack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01ca337d9fa391e4da960bc00f649039
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
